Fixed bug with save/restore stack
The restores and saves were unbalanced so saves would not necessarily restore to the correct spot. This was fixed by adding a clearSave method which is called when exiting a saved spot which pops the saved state from the stack but does not restore the position. This keeps the stack balanced which fixes the bug